requirements.txt
Flask==1.1.4
Werkzeug==1.0.1
python-dotenv==0.19.0
requests==2.31.0
urllib3==1.26.15
pyOpenSSL==23.2.0
certifi==2023.7.22
click==7.1.2
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==2.0.1 